requirements.txt
geopandas==0.9.0
numpy==1.19.2
pandas==1.2.4
streamlit==0.80.0
folium==0.12.1
dill==0.3.3
python_dateutil==2.8.1
